Database Manager changes

+++ Project-Lennon/Assets/Scripts/Database/DBManager.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public static class DBManager 
 {
     public static string username;
-    public static int score;
+    public static float score;
 
     public static bool LoggedIn { get { return username != null; } }
